personal-resume-assistant/pyproject.toml
[project]
name = "personal-resume-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.44.1",
    "openai>=1.102.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]
